+++ libs/ui/.storybook/preview.js
@@ -1,0 +1,1 @@
+import '!style-loader!css-loader!sass-loader!../../../apps/client/src/styles.scss';