+++ libs/ui/src/lib/value/value.component.ts
@@ -2,8 +2,7 @@ import {
   ChangeDetectionStrategy,
   Component,
   Input,
-  OnChanges,
-  OnInit
+  OnChanges
 } from '@angular/core';
 import { DEFAULT_DATE_FORMAT } from '@ghostfolio/common/config';
 import { format, isDate } from 'date-fns';
@@ -15,28 +14,26 @@ import { isNumber } from 'lodash';
   templateUrl: './value.component.html',
   styleUrls: ['./value.component.scss']
 })
-export class ValueComponent implements OnChanges, OnInit {
-  @Input() colorizeSign: boolean;
-  @Input() currency: string;
-  @Input() isCurrency: boolean;
-  @Input() isInteger: boolean;
-  @Input() isPercent: boolean;
-  @Input() label: string;
-  @Input() locale: string;
-  @Input() position: string;
-  @Input() size: string;
-  @Input() value: number | string;
+export class ValueComponent implements OnChanges {
+  @Input() colorizeSign = false;
+  @Input() currency = '';
+  @Input() isCurrency = false;
+  @Input() isInteger = false;
+  @Input() isPercent = false;
+  @Input() label = '';
+  @Input() locale = '';
+  @Input() position = '';
+  @Input() size = '';
+  @Input() value: number | string = '';
 
-  public absoluteValue: number;
-  public formattedDate: string;
-  public formattedValue: string;
-  public isDate: boolean;
-  public isNumber: boolean;
+  public absoluteValue = 0;
+  public formattedDate = '';
+  public formattedValue = '';
+  public isDate = false;
+  public isNumber = false;
   public useAbsoluteValue = false;
 
   public constructor() {}
-
-  public ngOnInit() {}
 
   public ngOnChanges() {
     if (this.value || this.value === 0) {